+++ cmd/bnscli/main.go
@@ -51,7 +51,6 @@ var commands = map[string]func(input io.Reader, output io.Writer, args []string)
 	"set-validators":            cmdSetValidators,
 	"sign":                      cmdSignTransaction,
 	"submit":                    cmdSubmitTransaction,
-	"tally":                     cmdTally,
 	"text-resolution":           cmdTextResolution,
 	"update-electorate":         cmdUpdateElectorate,
 	"update-election-rule":      cmdUpdateElectionRule,